feat(campaign): refactor CampaignPanel and ChatBot components for improved structure and naming consistency
