+++ client/src/components/campaign/CampaignPanel.tsx
